Refactor messages, improve mobile accessibility - Split messages into separate, reusable components - Move messages to below video on mobile - Size player to 16x9 on mobile - Disable send button if no message - Add icon to send button - Switch message field to single-line for compactness
